requirements.txt
celery==4.3.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.2
djangorestframework==3.7.7
djangorestframework-simplejwt==4.8.0
django-cors-headers==2.4.0
django-filter==1.1.0
django-postgres-extra==1.21a8
django-rest-swagger==2.1.2
idna==2.7
Markdown==2.6.11
psycopg2-binary==2.7.4
pytz==2018.3
typing==3.6.4
urllib3==1.23
requests==2.19.1
redis==3.2.1
sam-alignment-reconstructor==0.0.5
supervisor==4.2.1
mock==3.0.5
coveralls==2.2.0
testing.common.database==2.0.3
testing.postgresql==1.3.0
pylint==2.4.4
pylint-django==2.3.0
pyjwt[crypto]>=1.5.2